Flatten jenkins vars for overloading
This change is proposed in order to have a more fine grain control over
Jenkins install versions. With release of Jenkins 2.0 the current role
fails to install. This change allows for easy overriding of jenkins
variables in order to point to different mirrors.

Changes:
- Flatten the nested vars file
- Add jenkins namespacing prefixes.
- Abstracted out the debian apt_key and redhat repo and key urls in
order to allow for variable overloading.